feat: ハンバーガーメニュー内ユーザー機能リンクを動的に生成するよう変更
